Update criptografia e controllers
